errors

+++ server/src/main/java/server/websocket/WebSocketHandler.java
@@ -18,6 +18,7 @@ import service.LoginService;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.Objects;
 import java.util.Timer;
 
 
@@ -57,63 +58,109 @@ public class WebSocketHandler {
         int gameID = command.getGameID();
         String authToken = command.getAuthString();
         ChessGame.TeamColor playerColor = command.getPlayerColor();
-        ChessGame game = JoinGameService.getGame(gameID).getGame();
-        try {
-            connections.sendToClient(authToken, new LoadGameMessage(game));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+
+        // Test case: Join Player Bad GameID
+        if (!JoinGameService.gameExists(gameID)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid game ID."));
+            return;
         }
 
+        // Test case: Join Player Bad AuthToken
+        if (!JoinGameService.isValidAuthToken(authToken)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid authentication token."));
+            return;
+        }
+
+        // Test case: Join Player Wrong Team
+        if (playerColor != ChessGame.TeamColor.WHITE && playerColor != ChessGame.TeamColor.BLACK) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid team color."));
+            return;
+        }
+
+        // Test case: Join Player Empty Team
+        if (playerColor == null) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Team color cannot be empty."));
+            return;
+        }
+
+        ChessGame game = JoinGameService.getGame(gameID).getGame();
         connections.add(authToken, session);
+        connections.sendToClient(authToken, new LoadGameMessage(game));
         String username = LoginService.getUsername(authToken);
         var message = String.format("%s joined as a player", username);
         var notification = new NotificationMessage(message);
         connections.broadcast(authToken, notification);
     }
 
-    private void joinObserver(JoinObserverCommand command, Session session) throws DataAccessException, SQLException {
+    private void joinObserver(JoinObserverCommand command, Session session) throws DataAccessException, SQLException, IOException {
         int gameID = command.getGameID();
         String authToken = command.getAuthString();
-        String username = LoginService.getUsername(authToken);
-        ChessGame game = JoinGameService.getGame(gameID).getGame();
-        try {
-            connections.sendToClient(authToken, new LoadGameMessage(game));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+
+        // Test case: Join Observer Bad GameID
+        if (!JoinGameService.gameExists(gameID)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid game ID."));
+            return;
         }
 
+        // Test case: Join Observe Bad AuthToken
+        if (!JoinGameService.isValidAuthToken(authToken)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid authentication token."));
+            return;
+        }
+
+        String username = LoginService.getUsername(authToken);
+        ChessGame game = JoinGameService.getGame(gameID).getGame();
+        connections.sendToClient(authToken, new LoadGameMessage(game));
         connections.add(authToken, session);
         var message = String.format("%s joined as an observer", username);
         var notification = new NotificationMessage(message);
-        try {
-            connections.broadcast(authToken, notification);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
+        connections.broadcast(authToken, notification);
     }
 
-    private void makeMove(MakeMoveCommand command, Session session) throws SQLException, DataAccessException {
+    private void makeMove(MakeMoveCommand command, Session session) throws SQLException, DataAccessException, IOException {
         int gameID = command.getGameID();
         ChessMove move = command.getMove();
         String authToken = command.getAuthString();
+
+        // Test case: Make Invalid Move
+        if (move == null) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid move."));
+            return;
+        }
+
         ChessGame game = JoinGameService.getGame(gameID).getGame();
+
+        // Test case: Make Move Wrong Turn
+        if (!game.isCorrectTurn(move)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: It's not your turn."));
+            return;
+        }
+
+        String username = LoginService.getUsername(authToken);
+        ChessGame.TeamColor moveColor = game.getBoard().getPiece(move.getStartPosition()).getTeamColor();
+        String moveUsername;
+        if (moveColor == ChessGame.TeamColor.BLACK) {
+            moveUsername = JoinGameService.getGame(gameID).getBlackUsername();
+        } else {
+            moveUsername = JoinGameService.getGame(gameID).getWhiteUsername();
+        }
+        // Test case: Make Move for Opponent
+        if (!Objects.equals(moveUsername, username)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: You cannot make a move for the opponent."));
+            return;
+        }
 
         try {
             game.makeMove(move);
         } catch (InvalidMoveException e) {
             throw new RuntimeException(e);
         }
-
-
         LoadGameMessage updatedGameMessage = new LoadGameMessage(game);
         try {
             connections.broadcastAll(authToken, updatedGameMessage);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-
-        String username = LoginService.getUsername(authToken);
         String moveDescription = getMoveDescription(move, game);
         var notification = new NotificationMessage(username + " made a move: " + moveDescription);
         try {
@@ -121,7 +168,6 @@ public class WebSocketHandler {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-
         GameData gameData = JoinGameService.getGame(gameID);
         ChessGame.TeamColor currentUserColor = (username.equals(gameData.getWhiteUsername())) ? ChessGame.TeamColor.WHITE : ChessGame.TeamColor.BLACK;
         ChessGame.TeamColor opponentColor = (currentUserColor == ChessGame.TeamColor.WHITE) ? ChessGame.TeamColor.BLACK : ChessGame.TeamColor.WHITE;
@@ -129,9 +175,46 @@ public class WebSocketHandler {
         checkStatus(opponentColor, game, opponentUsername, authToken, gameID);
     }
 
-    private void leave(LeaveCommand command, Session session) throws DataAccessException {
+    private void resign(ResignCommand command, Session session) throws DataAccessException, SQLException, IOException {
         int gameID = command.getGameID();
         String authToken = command.getAuthString();
+
+        // Test case: invalidResignObserver
+        if (JoinGameService.isObserver(authToken, gameID)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Observers cannot resign."));
+            return;
+        }
+
+        // Test case: invalidResignGameOver
+        if (JoinGameService.isGameOver(gameID)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Cannot resign after the game is over."));
+            return;
+        }
+
+        endGame(gameID, authToken);
+        String username = LoginService.getUsername(authToken);
+        var message = String.format("%s resigned from game", username);
+        var notification = new NotificationMessage(message);
+        connections.broadcast(authToken, notification);
+    }
+
+
+    private void leave(LeaveCommand command, Session session) throws DataAccessException, SQLException, IOException {
+        int gameID = command.getGameID();
+        String authToken = command.getAuthString();
+
+        // Test case: Leave Bad GameID
+        if (!JoinGameService.gameExists(gameID)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid game ID."));
+            return;
+        }
+
+        // Test case: Leave Bad AuthToken
+        if (!JoinGameService.isValidAuthToken(authToken)) {
+            connections.sendToClient(authToken, new ErrorMessage("Error: Invalid authentication token."));
+            return;
+        }
+
         try {
             JoinGameService.removeFromGame(authToken, gameID);
         } catch (AuthenticationException e) {
@@ -145,27 +228,6 @@ public class WebSocketHandler {
         connections.remove(authToken);
         String username = LoginService.getUsername(authToken);
         var message = String.format("%s left game", username);
-        var notification = new NotificationMessage(message);
-        try {
-            connections.broadcast(authToken, notification);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void resign(ResignCommand command, Session session) throws DataAccessException, SQLException {
-        int gameID = command.getGameID();
-        String authToken = command.getAuthString();
-        ChessGame game = JoinGameService.getGame(gameID).getGame();
-
-        try {
-            endGame(gameID, authToken);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-
-        String username = LoginService.getUsername(authToken);
-        var message = String.format("%s resigned from game", username);
         var notification = new NotificationMessage(message);
         try {
             connections.broadcast(authToken, notification);
@@ -194,21 +256,8 @@ public class WebSocketHandler {
         boolean isInCheckmate = game.isInCheckmate(teamColor);
         boolean isInStalemate = game.isInStalemate(teamColor);
 
-        if (game.findKing(teamColor, game.getBoard()) == null) {
-            var notification = new NotificationMessage(username + "lost");
-            try {
-                connections.broadcastAll(authToken, notification);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-
-            try {
-                endGame(gameID, authToken);
-            } catch (SQLException e) {
-                throw new RuntimeException(e);
-            }
-        } else if (isInCheckmate) {
-            var notification = new NotificationMessage(username + " is in checkmate!");
+        if (isInCheckmate) {
+            var notification = new NotificationMessage(username + " is in checkmate! Game over.");
             try {
                 connections.broadcastAll(authToken, notification);
             } catch (IOException e) {
